fix(docs): Updated development documentation.
chore(docs): updated examples with explainations.

chore(docs): updated examples with explainations.

chore(docs): Added async example.

chore(docs): reverted yarn.lock file.

chore(docs): Updated with fake forms.

chore(docs): Updated with fake forms.

chore(docs): Updated with review comments.

Review updates.

Fixed bad link.
