+++ packages/module/patternfly-docs/content/examples/Advanced.tsx
@@ -20,7 +20,7 @@ export const AdvancedExample: React.FunctionComponent = () => {
             console.log (`Email Address: ${email}`);
             return false;
         }}
-        onOpenSupportCase='http://www.redhat.com'
+        onOpenSupportCase='https://pf-user-feedback-extension-form-demos.surge.sh/requestSupport.html'
         onReportABug={(email:string, bug:string)=>{
             console.log (`Email Address: ${email} Bug: ${bug}`)
             return true;}}
