+++ packages/module/patternfly-docs/content/examples/forms/requestSupport.html
@@ -1,0 +1,86 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <link rel="stylesheet" href="https://unpkg.com/@patternfly/patternfly/patternfly.min.css">
+</head>
+
+<body style="margin: 4rem; height: 80vh;">
+    <div class="pf-l-stack pf-m-gutter">
+        <div class="pf-l-stack__item">
+            <div class="pf-c-title pf-m-2xl">Request Support Example </div>
+        </div>
+        <div class="pf-l-stack__item">
+            <form class="pf-c-form pf-m-horizontal" novalidate>
+                <div class="pf-c-form__group">
+                    <div class="pf-c-form__group-label"><label class="pf-c-form__label" for="form-horizontal-name">
+                            <span class="pf-c-form__label-text">Email</span>&nbsp;<span class="pf-c-form__label-required"
+                                aria-hidden="true">&#42;</span></label>
+                    </div>
+                    <div class="pf-c-form__group-control">
+                        <input class="pf-c-form-control" type="text" id="form-horizontal-name"
+                            name="form-horizontal-name" required />
+                    </div>
+                </div>
+                <div class="pf-c-form__group">
+                    <div class="pf-c-form__group-label"><label class="pf-c-form__label" for="form-horizontal-info">
+                            <span class="pf-c-form__label-text">Issue</span></label>&nbsp;<span
+                            class="pf-c-form__group-label-help" aria-label="More information for information field"
+                            aria-describedby="form-horizontal-info" role="button" type="button" tabindex="0"><i
+                                class="pficon pf-icon-help" aria-hidden="true"></i></span>
+                    </div>
+                    <div class="pf-c-form__group-control">
+                        <textarea class="pf-c-form-control" type="text" id="form-horizontal-info"
+                            name="form-horizontal-info" aria-label="Textarea example"></textarea>
+                    </div>
+                </div>
+                <div class="pf-c-form__group" role="group" aria-labelledby="form-horizontal-checkbox-legend">
+                    <div class="pf-c-form__group-label pf-m-no-padding-top" id="form-horizontal-checkbox-legend"><span
+                            class="pf-c-form__label">
+                            <span class="pf-c-form__label-text">Department</span></span>&nbsp;<span
+                            class="pf-c-form__group-label-help" aria-label="More information for label field"
+                            aria-describedby="form-horizontal-checkbox-legend" role="button" type="button"
+                            tabindex="0"><i class="pficon pf-icon-help" aria-hidden="true"></i></span>
+                    </div>
+                    <div class="pf-c-form__group-control pf-m-stack">
+                        <div class="pf-c-check">
+                            <input class="pf-c-check__input" type="checkbox" id="form-horizontal-checkbox"
+                                name="form-horizontal-checkbox" />
+
+                            <label class="pf-c-check__label" for="form-horizontal-checkbox">Administration</label>
+                        </div>
+                        <div class="pf-c-check">
+                            <input class="pf-c-check__input" type="checkbox" id="form-horizontal-checkbox2"
+                                name="form-horizontal-checkbox2" />
+
+                            <label class="pf-c-check__label" for="form-horizontal-checkbox2">Engineering</label>
+                        </div>
+                        <div class="pf-c-check">
+                            <input class="pf-c-check__input" type="checkbox" id="form-horizontal-checkbox2"
+                                name="form-horizontal-checkbox2" />
+
+                            <label class="pf-c-check__label" for="form-horizontal-checkbox2">IT</label>
+                        </div>
+                        <div class="pf-c-check">
+                            <input class="pf-c-check__input" type="checkbox" id="form-horizontal-checkbox2"
+                                name="form-horizontal-checkbox2" />
+
+                            <label class="pf-c-check__label" for="form-horizontal-checkbox2">HR</label>
+                        </div>
+                        <div class="pf-c-check">
+                            <input class="pf-c-check__input" type="checkbox" id="form-horizontal-checkbox2"
+                                name="form-horizontal-checkbox2" />
+
+                            <label class="pf-c-check__label" for="form-horizontal-checkbox2">Security</label>
+                        </div>
+                    </div>
+                </div>
+            </form>
+        </div>
+        <div class="pf-l-stack__item">
+            <button class="pf-c-button pf-m-primary" style="width: fit-content;" type="button">Submit</button>
+            <button class="pf-c-button pf-m-secondary" style="width: fit-content" type="button">Cancel</button>
+        </div>
+    </div>
+
+</html>
