+++ packages/module/src/Feedback/FeedbackError.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button, Text, TextContent, TextVariants } from '@patternfly/react-core';
-import CheckIcon from '@patternfly/react-icons/dist/esm/icons/check-icon';
+import ErrorIcon from '@patternfly/react-icons/dist/esm/icons/error-circle-o-icon';
 import { LocaleContext } from '../context/LocaleContext';
 
 import './Feedback.scss';
@@ -13,7 +13,7 @@ const FeedbackError = ({ onCloseModal }: FeedbackErrorProps) => {
   const intl = React.useContext(LocaleContext);
   return (
     <div className="chr-c-feedback-success-content">
-      <CheckIcon size="md" color="var(--pf-global--success-color--100)" className="pf-u-mx-auto" />
+      <ErrorIcon size="md" color="var(--pf-global--danger-color--100)" className="pf-u-mx-auto" />
       <TextContent>
         <Text component={TextVariants.h1}>{intl.somethingWentWrong}</Text>
         <Text>
